move report to performance
