+++ athena/entrypoints/backtest.py
@@ -9,7 +9,7 @@ from athena.tradingtools.backtesting import (
     BacktestConfig,
     get_trades_from_strategy_and_fluctuations,
 )
-from athena.tradingtools.performance_report import build_and_save_trading_report
+from athena.tradingtools.performance.report import build_and_save_trading_report
 from athena.tradingtools.strategies import init_strategy
